Camera Setup
- Fixed code for camera setup.
- Modified code structure.
