+++ lib/heart_rate_detector/ui/state/heart_rate_state.dart
@@ -1,0 +1,3 @@
+import 'package:flutter/material.dart';
+
+class HeartRateState extends ChangeNotifier {}
