+++ lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:camera/camera.dart';
 import 'package:flutter/material.dart';
+import 'package:hear_rate_detector/heart_rate_detector/ui/screen/heart_rate_detector.dart';
 import 'package:permission_handler/permission_handler.dart';
 
 void main() {
@@ -74,39 +75,13 @@ class _MyHomePageState extends State<MyHomePage> {
 
   @override
   Widget build(BuildContext context) {
-    // This method is rerun every time setState is called, for instance as done
-    // by the _incrementCounter method above.
-    //
-    // The Flutter framework has been optimized to make rerunning build methods
-    // fast, so that you can just rebuild anything that needs updating rather
-    // than having to individually change instances of widgets.
     return Scaffold(
       appBar: AppBar(
-        // TRY THIS: Try changing the color here to a specific color (to
-        // Colors.amber, perhaps?) and trigger a hot reload to see the AppBar
-        // change color while the other colors stay the same.
         backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-        // Here we take the value from the MyHomePage object that was created by
-        // the App.build method, and use it to set our appbar title.
         title: Text(widget.title),
       ),
       body: Center(
-        // Center is a layout widget. It takes a single child and positions it
-        // in the middle of the parent.
         child: Column(
-          // Column is also a layout widget. It takes a list of children and
-          // arranges them vertically. By default, it sizes itself to fit its
-          // children horizontally, and tries to be as tall as its parent.
-          //
-          // Column has various properties to control how it sizes itself and
-          // how it positions its children. Here we use mainAxisAlignment to
-          // center the children vertically; the main axis here is the vertical
-          // axis because Columns are vertical (the cross axis would be
-          // horizontal).
-          //
-          // TRY THIS: Invoke "debug painting" (choose the "Toggle Debug Paint"
-          // action in the IDE, or press "p" in the console), to see the
-          // wireframe for each widget.
           mainAxisAlignment: MainAxisAlignment.center,
           children: <Widget>[
             const Text(
@@ -133,96 +108,12 @@ class _MyHomePageState extends State<MyHomePage> {
                 Navigator.push(
                     context,
                     MaterialPageRoute(
-                      builder: (context) => const DetectionScreen(),
+                      builder: (context) => const HeartRateDetector(),
                     ));
               },
         tooltip: 'Increment',
         child: const Icon(Icons.camera),
-      ), // This trailing comma makes auto-formatting nicer for build methods.
-    );
-  }
-}
-
-class DetectionScreen extends StatefulWidget {
-  const DetectionScreen({super.key});
-
-  @override
-  State<DetectionScreen> createState() => _DetectionScreenState();
-}
-
-class _DetectionScreenState extends State<DetectionScreen> {
-  @override
-  Widget build(BuildContext context) {
-    return const Scaffold(
-      body: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          Center(child: CameraWidget()),
-        ],
       ),
     );
-  }
-}
-
-class CameraWidget extends StatefulWidget {
-  const CameraWidget({super.key});
-
-  @override
-  State<CameraWidget> createState() => _CameraWidgetState();
-}
-
-class _CameraWidgetState extends State<CameraWidget> {
-  CameraPreview? _cameraPreview;
-  CameraFactory cameraFactory = CameraFactory.i;
-
-  @override
-  void initState() {
-    cameraFactory.createCamera().then((cameraPreview) {
-      _cameraPreview = cameraPreview;
-      setState(() {});
-    });
-
-    super.initState();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return (_cameraPreview == null)
-        ? const SizedBox(height: 100, width: 100, child: Placeholder())
-        : SizedBox(
-            height: 100,
-            width: 100,
-            child: _cameraPreview,
-          );
-  }
-
-  @override
-  void dispose() {
-    cameraFactory.disposeCamera();
-    super.dispose();
-  }
-}
-
-class CameraFactory {
-  CameraFactory._();
-  static final i = CameraFactory._();
-
-  CameraController? _controller;
-
-  Future<CameraPreview> createCamera() async {
-    var cameraDescription = await availableCameras();
-
-    if (cameraDescription.isEmpty) {
-      throw Exception('No cameras found');
-    }
-
-    _controller =
-        CameraController(cameraDescription.first, ResolutionPreset.high);
-
-    return CameraPreview(_controller!);
-  }
-
-  void disposeCamera() {
-    _controller?.dispose();
   }
 }
